requirements.txt
argparse
dnspython
whoisit
